[PSym] refactor: separate explicit from symbolic (part 1)
