+++ Src/PRuntimes/PSymRuntime/src/main/java/psym/commandline/PSymOptions.java
@@ -4,13 +4,13 @@ import org.apache.commons.cli.*;
 import org.json.JSONArray;
 import org.json.JSONObject;
 import org.json.JSONTokener;
-import psym.runtime.scheduler.choiceorchestration.ChoiceLearningRewardMode;
-import psym.runtime.scheduler.choiceorchestration.ChoiceLearningStateMode;
-import psym.runtime.scheduler.choiceorchestration.ChoiceOrchestrationMode;
-import psym.runtime.scheduler.choiceorchestration.ChoiceOrchestratorEpsilonGreedy;
+import psym.runtime.scheduler.explicit.choiceorchestration.ChoiceLearningRewardMode;
+import psym.runtime.scheduler.explicit.choiceorchestration.ChoiceLearningStateMode;
+import psym.runtime.scheduler.explicit.choiceorchestration.ChoiceOrchestrationMode;
+import psym.runtime.scheduler.explicit.choiceorchestration.ChoiceOrchestratorEpsilonGreedy;
 import psym.runtime.scheduler.symmetry.SymmetryMode;
-import psym.runtime.scheduler.taskorchestration.TaskOrchestrationMode;
-import psym.runtime.scheduler.taskorchestration.TaskOrchestratorCoverageEpsilonGreedy;
+import psym.runtime.scheduler.explicit.taskorchestration.TaskOrchestrationMode;
+import psym.runtime.scheduler.explicit.taskorchestration.TaskOrchestratorCoverageEpsilonGreedy;
 import psym.utils.GlobalData;
 import psym.utils.StateCachingMode;
