+++ Src/PRuntimes/PSymRuntime/src/main/java/psym/runtime/scheduler/IterativeBoundedScheduler.java
@@ -6,9 +6,10 @@ import psym.commandline.PSymConfiguration;
 import psym.commandline.Program;
 import psym.runtime.logger.*;
 import psym.runtime.machine.Machine;
-import psym.runtime.scheduler.choiceorchestration.*;
+import psym.runtime.scheduler.explicit.BacktrackTask;
+import psym.runtime.scheduler.explicit.choiceorchestration.*;
 import psym.runtime.scheduler.symmetry.SymmetryMode;
-import psym.runtime.scheduler.taskorchestration.TaskOrchestrationMode;
+import psym.runtime.scheduler.explicit.taskorchestration.TaskOrchestrationMode;
 import psym.runtime.statistics.SearchStats;
 import psym.utils.GlobalData;
 import psym.utils.MemoryMonitor;
