+++ Src/PRuntimes/PSymRuntime/src/main/java/psym/runtime/scheduler/explicit/choiceorchestration/ChoiceLearningStateMode.java
@@ -1,4 +1,4 @@
-package psym.runtime.scheduler.choiceorchestration;
+package psym.runtime.scheduler.explicit.choiceorchestration;
 
 public enum ChoiceLearningStateMode {
     None,
