+++ Src/PRuntimes/PSymRuntime/src/main/java/psym/runtime/scheduler/explicit/choiceorchestration/ChoiceOrchestratorQLearning.java
@@ -1,5 +1,6 @@
-package psym.runtime.scheduler.choiceorchestration;
+package psym.runtime.scheduler.explicit.choiceorchestration;
 
+import psym.utils.GlobalData;
 import psym.utils.RandomNumberGenerator;
 import psym.valuesummary.ValueSummary;
 
@@ -7,12 +8,12 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Random;
 
-public class ChoiceOrchestratorRandom implements ChoiceOrchestrator {
-
-    public ChoiceOrchestratorRandom() {
+public class ChoiceOrchestratorQLearning implements ChoiceOrchestrator {
+    public ChoiceOrchestratorQLearning() {
     }
 
     public void reorderChoices(List<ValueSummary> choices, boolean isData) {
         Collections.shuffle(choices, new Random(RandomNumberGenerator.getInstance().getRandomLong()));
+        Collections.sort(choices, GlobalData.getChoiceLearningStats().getChoiceComparator());
     }
 }
