+++ Src/PRuntimes/PSymRuntime/src/main/java/psym/runtime/scheduler/explicit/taskorchestration/TaskOrchestratorCoverageEpsilonGreedy.java
@@ -1,7 +1,7 @@
-package psym.runtime.scheduler.taskorchestration;
+package psym.runtime.scheduler.explicit.taskorchestration;
 
 import lombok.Setter;
-import psym.runtime.scheduler.BacktrackTask;
+import psym.runtime.scheduler.explicit.BacktrackTask;
 import psym.utils.RandomNumberGenerator;
 
 public class TaskOrchestratorCoverageEpsilonGreedy implements TaskOrchestrator {
